Changed data access logic
