+++ WebShop/src/main/java/com/tms/webshop/servlets/user/NewAccountServlet.java
@@ -1,7 +1,7 @@
 package com.tms.webshop.servlets.user;
 
-import com.tms.webshop.model.UserType;
 import com.tms.webshop.dao.UserDao;
+import com.tms.webshop.model.UserType;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
@@ -26,11 +26,11 @@ public class NewAccountServlet extends HttpServlet {
         String passwordRepeat = request.getParameter("passwordRepeat");
         LocalDate birthday = LocalDate.parse(request.getParameter("birthday"));
 
-        UserDao userDAO = new UserDao();
+        UserDao userDao = (UserDao) request.getServletContext().getAttribute(UserDao.CONTEXT_NAME);
 
-        if (isLoginValid(login, userDAO) || isNameValid(name) || isNameValid(surname) || isEmailValid(email) ||
+        if (isLoginValid(login, userDao) || isNameValid(name) || isNameValid(surname) || isEmailValid(email) ||
                 password.equals(passwordRepeat) || isBirthdayValid(birthday)) {
-            userDAO.addUser(login, password, UserType.CLIENT, name, surname, email, birthday);
+            userDao.addUser(login, password, UserType.CLIENT, name, surname, email, birthday);
         } else {
             response.setStatus(400);
         }
@@ -48,7 +48,7 @@ public class NewAccountServlet extends HttpServlet {
         return !name.isEmpty() && Pattern.compile("[A-Z][a-z]*").matcher(name).matches();
     }
 
-    public boolean isLoginValid(String login, UserDao userDAO) {
-        return !login.isEmpty() && !userDAO.loginInUse(login);
+    public boolean isLoginValid(String login, UserDao userDao) {
+        return !login.isEmpty() && !userDao.loginInUse(login);
     }
 }
