+++ WebShop/src/main/java/com/tms/webshop/servlets/user/admin/AdminAddCategoryServlet.java
@@ -23,10 +23,10 @@ public class AdminAddCategoryServlet extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         InputStream fileStream = request.getPart("image").getInputStream();
 
-        ImageDao imageDAO = new ImageDao();
-        imageDAO.addImage(request.getParameter("imageName"), fileStream);
+        ImageDao imageDao = (ImageDao) request.getServletContext().getAttribute(ImageDao.CONTEXT_NAME);
+        imageDao.addImage(request.getParameter("imageName"), fileStream);
 
-        CategoryDao categoryDAO = new CategoryDao();
-        categoryDAO.addCategory(request.getParameter("name"), request.getParameter("imageName"));
+        CategoryDao categoryDao = (CategoryDao) request.getServletContext().getAttribute(CategoryDao.CONTEXT_NAME);
+        categoryDao.addCategory(request.getParameter("name"), request.getParameter("imageName"));
     }
 }
